+++ SynthDicom.Tests/PackageListIsCorrectTests.cs
@@ -64,7 +64,7 @@ public sealed class PackageListIsCorrectTests
     {
         if (path != null)
         {
-            if (!Path.IsPathRooted(path)) path = Path.Combine(TestContext.CurrentContext.TestDirectory, path);
+            if (!Path.IsPathRooted(path)) path = Path.Join(TestContext.CurrentContext.TestDirectory, path);
             return new DirectoryInfo(path);
         }
         var root = new DirectoryInfo(TestContext.CurrentContext.TestDirectory);